fix/configure_snapraid: clear `SNAPPER_CONFIGS` in all config paths
`/etc/default/snapper` was previously the default, but in newer versions of snapper it seems to have moved to `/etc/sysconfig/snapper`. Sometimes it still looks at both config files, to it's critical to make this change in both places if they both exist.

Set `ignore_errors: true` to allow for missing files, as they may not both exist.